refactor: 🚜 Moved the passwordMatch function to be a in the utilities instead of a service.
